Added buttonCluster and object cluster creation in grid
Changes:

Added a new button called buttonCluster to create an object cluster on the grid.

Improved the way items' locations on the grid are displayed. The methods have been moved from the mainwindow.cpp file to the griditem.cpp file, along with small method declarations in the headers.

Observations:

Renamed the variable buttonA in the mainwindow.ui file to buttonPC.

Removed the functionality of a button in the userwindow.ui file that previously opened the settingswindow for testing purposes.